add some exceptions and id validtion on services

+++ src/main/java/com/mindhub/ms_order/services/Impl/OrderEntityServiceImpl.java
@@ -1,10 +1,12 @@
 package com.mindhub.ms_order.services.Impl;
 
 import com.mindhub.ms_order.dtos.OrderEntityDTO;
+import com.mindhub.ms_order.exceptions.NotFoundException;
 import com.mindhub.ms_order.mappers.OrderEntityMapper;
 import com.mindhub.ms_order.models.OrderEntity;
 import com.mindhub.ms_order.repositories.OrderEntityRepository;
 import com.mindhub.ms_order.services.OrderEntityService;
+import org.hibernate.annotations.NotFound;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -20,13 +22,21 @@ public class OrderEntityServiceImpl implements OrderEntityService {
     OrderEntityMapper orderEntityMapper;
 
     @Override
-    public OrderEntityDTO getOrder(Long id) throws Exception {
-        return orderEntityMapper.orderToDTO(findById(id));
+    public OrderEntityDTO getOrder(Long id) throws NotFoundException {
+        if(existsById(id)) {
+            return orderEntityMapper.orderToDTO(findById(id));
+        } else {
+            throw new NotFoundException("Not found.");
+        }
     }
 
     @Override
-    public OrderEntity getOrderEntity(Long id) throws Exception {
-        return findById(id);
+    public OrderEntity getOrderEntity(Long id) throws NotFoundException {
+        if(existsById(id)) {
+            return findById(id);
+        } else {
+            throw new NotFoundException("Not found.");
+        }
     }
 
     @Override
@@ -40,19 +50,27 @@ public class OrderEntityServiceImpl implements OrderEntityService {
     }
 
     @Override
-    public OrderEntity updateOrder(Long id, OrderEntityDTO updatedOrder) throws Exception {
-        OrderEntity orderToUpdate = findById(id);
-        return save(orderEntityMapper.updateOrderToEntity(orderToUpdate, updatedOrder));
+    public OrderEntity updateOrder(Long id, OrderEntityDTO updatedOrder) throws NotFoundException {
+        if(existsById(id)) {
+            OrderEntity orderToUpdate = findById(id);
+            return save(orderEntityMapper.updateOrderToEntity(orderToUpdate, updatedOrder));
+        } else {
+            throw new NotFoundException("Not found.");
+        }
     }
 
     @Override
     public void deleteOrder(Long id) {
-        deleteById(id);
+        if(existsById(id)) {
+            deleteById(id);
+        } else {
+            throw new NotFoundException("Not found.");
+        }
     }
 
     @Override
-    public OrderEntity findById(Long id) throws Exception {
-        return (orderEntityRepository.findById(id).orElseThrow(() -> new Exception("Not Found.")));
+    public OrderEntity findById(Long id) throws NotFoundException {
+        return (orderEntityRepository.findById(id).orElseThrow(() -> new NotFoundException("Not found.")));
     }
 
     @Override
@@ -61,12 +79,17 @@ public class OrderEntityServiceImpl implements OrderEntityService {
     }
 
     @Override
-    public void deleteById(long id) {
+    public void deleteById(Long id) {
         orderEntityRepository.deleteById(id);
     }
 
     @Override
     public OrderEntity save(OrderEntity order) {
         return orderEntityRepository.save(order);
+    }
+
+    @Override
+    public Boolean existsById(Long id) {
+        return orderEntityRepository.existsById(id);
     }
 }
